move css injection to embeddable-ui-embeddable
