+++ lifecycle.config.ts
@@ -1,31 +1,6 @@
-const generateCssVariables = (variables: Record<string, string>) => {
-  let textContent = '';
-  Object.keys(variables).forEach((key) => {
-    const value = variables[key];
-    textContent += `${key}: ${value};\n`;
-  });
-  return textContent;
-};
-
-// TODO: check the possibility of injecting via CSS
-export const injectCssVariables = (styles: Record<string, string>) => {
-  const css = `:root {\n${generateCssVariables(styles)}}`;
-  const styleId = 'embeddable-style';
-  let styleEl = document.getElementById(styleId) as HTMLStyleElement | null;
-
-  if (styleEl) {
-    // overwrite the old vars
-    styleEl.textContent = css;
-  } else {
-    styleEl = document.createElement('style');
-    styleEl.id = styleId;
-    styleEl.textContent = css;
-    document.head.appendChild(styleEl);
-  }
-
-  return () => styleEl?.remove();
-};
+import { injectCssVariables } from './src/remarkable-ui-embeddables/theme/styles/stytles.utils';
+import { Theme } from './src/remarkable-ui-embeddables/theme/theme.types';
 
 export default {
-  onThemeUpdated: ({ styles }) => injectCssVariables(styles),
+  onThemeUpdated: (theme: Theme) => injectCssVariables(theme.styles),
 };
