+++ src/remarkable-ui/styles/styles.utils.ts
@@ -1,33 +1,5 @@
 import { StylesKeys } from './styles.constants';
 
-const generateCssVariables = (variables: Record<string, string>) => {
-  let textContent = '';
-  Object.keys(variables).forEach((key) => {
-    const value = variables[key];
-    textContent += `${key}: ${value};\n`;
-  });
-  return textContent;
-};
-
-// TODO: check the possibility of injecting via CSS
-export const injectCssVariables = (styles: Record<string, string>) => {
-  const css = `:root {\n${generateCssVariables(styles)}}`;
-  const styleId = 'embeddable-style';
-  let styleEl = document.getElementById(styleId) as HTMLStyleElement | null;
-
-  if (styleEl) {
-    // overwrite the old vars
-    styleEl.textContent = css;
-  } else {
-    styleEl = document.createElement('style');
-    styleEl.id = styleId;
-    styleEl.textContent = css;
-    document.head.appendChild(styleEl);
-  }
-
-  return () => styleEl?.remove();
-};
-
 /**
  * Attempts to interpret a CSS variable as a number (for lengths)
  * or a string (for colors, etc.).
